Improve alignment of output (#6)
This change puts the categories output in nice straight columns by
omitting the full-precision number of bytes and printing the other
fields using a fixed number of characters. If the percentage goes to
100% or a quantity hits, say, 1023.0 MB, then the columns will not be
perfect, but usually they are.

The percentage column becomes particularly handy because for the values
that aren't sorted it can be used to easily find the largest values, by
scanning down the column.

This fixes #3.